\name{alliances1943}
\docType{data}
\alias{alilances1943}
\title{Alliance Portfolios of European Major Powers and U.S., 1947}
\description{
  This dataset contains the alliance portfolios of the USA, France,
  Germany and Russia in 1947, along with the nations' proportion of system capabilities. 
}
\usage{alliances1947}
\format{
  A dataframe with 6 columns and 26 rows
  \tabular{rlll}{
    [,1] \tab Nation  \tab factor \tab Country Code \cr
    [,2] \tab SysCap  \tab numeric \tab System Capabilities \cr
    [,3] \tab USA    \tab integer \tab Alliances with USA (0-3) \cr
    [,4] \tab UK    \tab integer \tab Alliances with the United Kingdom (0-3) \cr
    [,5] \tab FRN   \tab integer \tab Alliances with France (0-3) \cr
    [,6] \tab RUS     \tab integer \tab Alliances with the Soviet Union (0-3)
  }

}
\source{Signorino and Ritter (1999), Table 7(a), p. 136}
\references{
  Signorino, C. S. and Ritter, J.M. (1999)
  Tau-b or Not Tau-b, \emph{International Studies Quarterly}, \bold{43}, 115--144.
}
\keyword{datasets}
